AetherSign/requirements_v0.txt
numpy>=1.24
PyYAML>=6.0
tensorflow>=2.13
torch>=2.0
